Update GPU Pixel_CSSFilterEffects parameters
Updates the matching algorithm parameters for the GPU
Pixel_CSSFilterEffects tests, as the old parameters were too strict to
properly handle certain noisy configurations.

The new algorithm is a pure fuzzy matching algorithm instead of a
Sobel filter algorithm, as the large number of gradients and blurs in
the image were not playing nicely with the Sobel filter.

TBR=ynovikov@chromium.org

Bug: 1087740
Change-Id: I50d9cf37dde97d215af8436a801cf0e5555bbc4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222860
Reviewed-by: Brian Sheedy <bsheedy@chromium.org>
Reviewed-by: Yuly Novikov <ynovikov@chromium.org>
Commit-Queue: Brian Sheedy <bsheedy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773404}