weblayer: Avoid resize to 0 on rotation
When fragment is retained by activity is recreated on rotation,
ContentViewRenderView is recreated and the same WebContents is moved to
the new CVRV. CVRV.setWebContents is called on the new CVRV before its
first layout, which means it still has size 0x0. This causes a momentary
resize of the WebContents to 0x0 before being resized again to the
correct size. On complex websites, resize and relayout is expensive,
and this extra step does add up to the latency on rotation.

Just check for 0x0 in this case, and skip the resize here.

Bug: 1084500
Change-Id: Idea96b8df53e29338af4c1d91851a69ddad8f45b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2213076
Reviewed-by: Scott Violet <sky@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773422}