[CrOS Settings] Fix "Device" search result
Previously, a search for "Device" did not actually show the "Device"
result. This was because there are many matches for this query, and
LocalSearchService would not return it as one of the top results. Since
we capped the query, we would never receive this result back.

Now, we query for up to 5x the max results, then wait until after the
returned results are sorted before truncating the list.

Bug: 1088058
Change-Id: I87d63ca27042fd9ce69142540562f40815a4686e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222533
Commit-Queue: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Auto-Submit: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773453}