Cleanup: Renamed AXSelectionBehavior members to clarify that it works both on the DOM and the AX trees
AXSelectionBehavior is used to determine whether a DOM selection will be extended or shrunk if
its endpoints refer to an invalid position in the accessibility tree.
However, it is also used in methods such as AXSelection::Select and AXSelection::AsSelection to
specify the same concerning converting a selection range in the accessibility tree to a near equivalent range in the DOM tree.

AX-Relnotes: n/a.

R=dmazzoni@chromium.org

Change-Id: I689afda1c54628c9e622d8d95cc2fe7fb1d6cb9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219904
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773533}