Ensures that the anchor and focus of a selection is updated to an ignored node if needed
In https://crrev.com/c/2205400/ a fix was made to the
selection offsets that might be necessary whenever there are
nodes in the internal accessibility tree that are ignored, i.e.,
they are not exposed to the platform tree.

This patch improves on that work by ensuring that
when setting the selection, the anchor or focus nodes are appropriately
set to an ignored node, if either of the selection endpoints are
at a tree position before a node whose parent is ignored
in the platform tree but is unignored in the internal tree.

HTML:
<p>
  <div role="presentational">
    <span>hello</span>
  </div>
</p>

Internal tree:
++Paragraph
++++GenericContainer ignored
++++++StaticText hello

Platform tree:
++Paragraph
++++StaticText hello

Set Selection to the whole of the paragraph's text.
Selectin in platform tree:
Anchor=Paragraph, 0; Focus=Paragraph, 1

Selection in internal tree:
Anchor=GenericContainer, 0; focus=GenericContainer, 1

AX-Relnotes: n/a.

R=dmazzoni@chromium.org

Bug: 1082814
Change-Id: Ib1775eeae2c74c118d4084fe6c0c9b4a80152d78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2218323
Commit-Queue: Nektarios Paisios <nektar@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773534}