Add window segments concept to the content layer
This change adds a the window segments concept to the
SynchronizeVisualProperties cascade. A display feature splits a view by,
e.g., a fold or a mask - an accessor is added to RenderWidgetHostViewBase
to provide this information during the calculation of visual properties.
When the root render widget is split by the display feature, there are
multiple 'window segments' that web authors will be able to use to split
their view and have a logical separation of content (either through
JavaScript or CSS).

The root widget window segments are computed by the root
RenderWidgetHostImpl, based on the display feature on the
RenderWidgetHostView. The segments are then propagated to child local
frame roots via RenderFrameProxy and CrossProcessFrameConnector via the
new member on VisualProperties.

Follow-up changes will finish hooking this information up to Blink in
order to expose it to the web, and also integrate with the platform
specific APIs in order to get the display feature information from the
underlying operating system.

Explainers:
https://github.com/webscreens/window-segments/blob/master/EXPLAINER.md
https://github.com/MicrosoftEdge/MSEdgeExplainers/blob/master/Foldables/explainer.md

i2p:
https://groups.google.com/a/chromium.org/d/msg/blink-dev/_1_ljYsaqrU/qOoHJ7MhAgAJ
https://groups.google.com/a/chromium.org/d/msg/blink-dev/KsPRO3hkOGw/md4jgIAaAgAJ

Bug: 1039050 1039051
Change-Id: I6805cda5946bd5379ad6dcd68e092e5d7d9be42d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2205123
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773545}