[consent toggle] Register pref and add settings toggle.
This change does two things:
1) Create and register a pref that tracks whether Suggested Content is
   enabled
2) Add a new toggle to the privacy settings that controls this pref.

Since the Suggested Content feature itself is not implemented yet,
there is currently no callback for when the pref is toggled.

The URL in chrome/common/url_constants.cc is a placeholder for now and
will be replaced when the support team provides a link. The feature is
behind a chrome/finch flag.

Bug: 1079169
Change-Id: Ida3b210811aafc2b96a49df09822ebf6cdfac885
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2192534
Commit-Queue: Rachel Wong <wrong@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Thanh Nguyen <thanhdng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773593}