[Files App] Override open zip behavior with ZipNoNaCl flag
When selecting a zip file, make the open task mount the file with the
Zip Archiver if the FilesZipNoNaCl feature is enabled, or display an
error message saying that the new zip handling isn't implemented
otherwise.

A progress panel is used to display the error message: the
progressCenter instance of FileManager is added as a FileTasks
attribute.

browser_tests --gtest_filter=*zipFileOpenDownloads
browser_tests --gtest_filter=*zipFileCannotOpen*

Test: 
Bug: 912236
Change-Id: I25a29b7dd5bb25cb918b6ff98e08a7665251500b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2217940
Commit-Queue: Jeremie Boulic <jboulic@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: François Degros <fdegros@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773627}