SkiaRenderer/macOS: Implement basic paths for macOS
A handful of odds-and-ends here.

Add a mailbox argument to CALayerParams, as this is how we will
reference the shared image in SkiaRenderer (texture ID won't do).

Change the CALayerOverlaySharedStateto be thread-safe ref-counted, as it
will now be passed across threads (previously it was just serialized).

Populate the SkiaRenderer::ScheduleOverlays path for macOS. This will
need further work to take into account the unique lifetime behaviors
of IOSurfaces.

Populate SkiaOutputDeviceBufferQueue::ScheduleOverlays for macOS and
also Ozone.

Bug: 894929
Change-Id: I0ab0ce7ed1aacd24a00b3854dc332bab58c134cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222129
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773721}