SkiaRenderer/macOS: Use SwapCompletionResult for SwapCompletionCallback
The gl::GLSurface::SwapCompletionCallback function takes two arguments:
a gfx::SwapResult and gfx::GpuFence. Consolidate these into a single
SwapCompletionResult structure. This mirrors the behavior of
gl::GLSurface::PresentationCallback, which takes a
gfx::PresentationFeedback.

In a follow-on patch, we will add a gfx::CALayerParams argument to
this structure, which will be used to communicate information about
the CALayer tree on macOS.

Bug: 894929
Change-Id: I03eeec04a1b11f1bc8b99d6e426ed04897941061
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2223376
Reviewed-by: Jonathan Backer <backer@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773749}