Clean up single item folder after initial sync.
App list sync may have some edge cases that lead to the creation of
single item folder which is not legitimate for launcher. When user signs
in a new session, app list sync service will download all the sync data
from sync backend and merge them with the cached sync data in local
storage. We will detect the single item folder sync items and clean them
up right after the initial merging completes in AppListSyncableService.
Therefore, the single item folder should go away from the launcher UI
after user starts a new session.

This does not fix the root cause of issues that lead to the creation of
the single item folder, but makes it a little better with a remedy. We
will also fix root cause of the issue for the cases we can repro in
separate changelists.

Bug: 1082530
Change-Id: I95098e0c97eb3faaf8d90d6654d9f1ca8a60094c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219176
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Jenny Zhang <jennyz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773777}