Add flag to use LibAssistant Beta backend
This flag |LibAssistantBetaBackend| can be given as a command-line
flag, and will tell LibAssistant to use the BETA_DOGFOOD backend.
As a practical side-effect, this will show more useful information in
the logs, like the actual queries that were submitted.

Bug: b/157557676
Change-Id: I580461b649cdbae74f1d658329b1135d73823fdd
Tests: Launched Assistant with and without this flag.
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225260
Commit-Queue: Jeroen Dhollander <jeroendh@google.com>
Reviewed-by: Tao Wu <wutao@chromium.org>
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773787}