[OSL-RV] Temporary memory increase mitigation
This is a change that will hopefully help us understand the regression
and in the long run propose a reasonable trade off between memory
cost of view reuse, performance and functionality we get.

As an added bonus, this should make the view creation numbers match
those of a ListView more closely.

Bug: 1087720, 1087691
Change-Id: Iaaccb17f1eaa59058179e476ef303cfdaff68611
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225128
Commit-Queue: Filip Gorski <fgorski@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773788}