indexeddb: refactor special storage policy
Special storage policy is a dynamic, complicated, browser-only data
structure that IndexedDB needs to query on shutdown to determine whether
or not to delete particular session-only directories (or not delete
protected directories).  Because shutdown is time-sensitive, this
chooses to push updates to policies based on active origins (i.e.
origins that IDBFactories have been created for).

This is the IndexedDB equivalent of this local storage change:
https://chromium-review.googlesource.com/c/chromium/src/+/1945940

This is a reland (with better thread safety) of:
https://chromium-review.googlesource.com/c/chromium/src/+/2203146

Bug: 1015214
Change-Id: I1d62d5ed93e45a33a85d0935ad7f211428283fe4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216806
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Commit-Queue: enne <enne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773813}