Allow common words to be embedded in Target Embedding.
Some top domains use a common word for their website name. Office.com
and weather.com are examples of such domains. As a result, hostnames
embedding these common words might be falsely flagged. We allowlist
some common words in this CL that have shown to lead to false
positives. There are two types of common word allowlisting. One is
allowing the common word to be embedded with all TLDs and two is
allowing the common word to be embedded with TLDs other than the TLD of
the top domain. For example, weather is a common word for all TLDs.
As a result, neither of weather.com-foo.com, and weather.info-foo.com
would trigger this heuristic. However, office is a common word just for
TLDs other than ".com". So, office.info-foo.com won't trigger the
heuristic, but office.com-foo.com would trigger it.


Bug: 1087636
Change-Id: Id5b6bc5204254f45429a7d91690af192fcef81e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2224313
Commit-Queue: Behnood Momenzadeh <behnoodm@google.com>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773839}