Fix subframe history loads when both old and new FrameNavigationEntry is missing.
This CL fixes a regression introduced by the refactor in
https://chromium-review.googlesource.com/c/chromium/src/+/2181973.
Apparently, it's possible that a subframe may have neither the old nor
the new (target) FrameNavigationEntry during a history navigation.
Prior to the refactor above, we returned early and avoided scheduling
any loads for such a subframe.  After the refactor, we incorrectly
attempted to schedule a different-document load, which resulted in
calling CreateNavigationRequestFromEntry() with a null frame_entry,
which it didn't expect and crashed while deferencing it.

Bug: 1088354, 1088175
Change-Id: Iaaa9743b0e897eecb1aa0317fa20d8d042c878f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225267
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773863}