Add accessibility for Quick-Answers UserConsentView
Announce the user-consent notice for Quick Answers when screen-reader is
enabled. The accessibility name for the view is composed of all of its
text content concatenated, by default. The accessible name and
description are overridden for the view, and the announcement is made
when the user-consent view is displayed.

Bug: b:152059023
Test: Tested on Chrome OS Atlas test device.
Change-Id: I222b139263ea748293112c26df37c226b0e43634
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222126
Commit-Queue: Abhijeet Singh <siabhijeet@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773968}