[ScrollUnification] Fix scrolling without root layer
This CL specifically fixes behavior for the test
LayerTreeHostImplTest.ScrollBeforeRootLayerAttached as well as some
others that check the "viewport fallback" behavior when the scroll
unification flag is enabled.

This test was a regression test for a crash that was occurring when a
scroll occurred prior to a root layer being attached. In this case, we
hit test to a nullptr for the layer. Today, we drop the scroll in this
case. The initial scroll unification implementation inappropriately
requestes a main thread hit test.

The fix here replicates this logic in the kScrollUnification branch of
ScrollBegin so that behavior is unchanged. When a hit test doesn't hit
any layers, we currently fallback to using the inner viewport scroll
node. There are some cases where the inner viewport scroll node doesn't
exist:

 - UI Comopsitor never has viewport nodes
 - OOPIF never has viewport nodes
 - Have not committed from Blink yet

The key problem here was that the scroll node hit test returning nullptr
is actually two distinct cases that we weren't distinguishing:

 1. The compositor realizes it doesn't have enough information to
    perform the hit test (e.g. hit a non-fast scrolling region)
 2. The compositor is able to perform the hit test and there simply
    isn't a ScrollNode.

1 requires a main thread hit test where 2 must drop the scroll. The
HitTestScrollNode method is modified here to return the additional
information and makes this explicit.

Bug: 1047182,1086625
Change-Id: Id547e73dc307f0118b6c8cd3625a26661b57a45d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2222140
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#773985}